+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes will be documented in this file.
 
 ### Fixed
 
-* PAPI incorrect OrderAck from modify response when order was canceled (#603)
+* Incorrect OrderAck from modify response when order was canceled (#603)
 
 ## 1.1.5 &ndash; 2026-06-06